Add new trait for plugin loading
The current plugin loader trait defines some semantically wrong
"eventually..." methods and also functions as a factory. The new plugin
loading approach must just provide the plugin's fully qualified class
name. Construction of the plugins is then up to the consumer.
Plus, the new trait's class name follows our naming convetion for
traits, so that it is more clear what is enabled when using the trait.